+++ webpack.config.js
@@ -1,0 +1,25 @@
+const path = require('path');
+const TerserPlugin = require("terser-webpack-plugin");
+
+module.exports = {
+  devtool: "source-map",
+  entry: './src/index.js',
+  optimization: {
+    minimize: false,
+    minimizer: [
+      new TerserPlugin({
+        terserOptions: {
+          format: {
+            comments: false,
+          },
+        },
+        extractComments: false,
+      }),
+    ],
+  },
+  output: {
+    filename: 'main.js',
+    path: path.resolve(__dirname, 'dist'),
+  },
+};
+